rename: move headers under `include/autoware/ground_filter`
